Fixed setup.py so that it doesn't throw encoding error.
Without specifying the encoding of the README file, the open defaults to system config encoding, which doesn't have to be 'utf-8'. Opening the file in UTF-8 mode fixes that.

Traceback of the error thrown without the fix:

`Traceback (most recent call last):
  File "setup.py", line 156, in <module>
    README_rst = fd.read()
  File "C:\Users\n0272436\Downloads\WinPython-32bit-3.4.3.4\python-3.4.3\lib\encodings\cp1252.py", line 23,
    return codecs.charmap_decode(input,self.errors,decoding_table)[0]
UnicodeDecodeError: 'charmap' codec can't decode byte 0x8d in position 17233: character maps to <undefined>
`

The error in question is referred to, and discussed here:

http://stackoverflow.com/questions/25049962/is-encoding-is-an-invalid-keyword-error-inevitable-in-python-2-x