fix navigation for windows command shell
if other windows version, use pure navigation.

Future: detect if ANSI.sys loaded, because in that case, virtual terminal escape works too.

Reason: windows 64bit (with the exception of windows10) `cmd` doesn't allow the command shell to effectively support colorama or moves-via-write-to-stdout, as ANSI.sys doesn't exist and is therefore not loaded. Windows10 has a special terminal mode which can be engaged which allows move-via-write.