Assign to _instances and monitor under the class lock
This ensures that when two instances of tqdm are created simultaneously, there
will be no two WeakSet or two TMonitor objects.

This ensures that when the last instance is closed and a new one is created
simultaneously, the last instance will not delete the monitor that serves the
new instance.