Speed up tqdm.auto import when not in an IPython notebook
In tqdm.auto, look for `IPython` in `sys.modules` rather than
explicitly importing it. If we are in an IPython notebook, then
IPython will have already been imported. If we are not in an
IPython notebook, then importing IPython will take almost half a
second.

Here are import times on a 2018 MacBook Pro with an SSD.
Before this patch:

    $ time python -c 'from tqdm.auto import tqdm'

    real     0m0.474s
    user     0m0.278s
    sys      0m0.142s

After this patch:

    $ time python -c 'from tqdm.auto import tqdm'

    real    0m0.069s
    user    0m0.041s
    sys     0m0.020s

Fixes #709.