Added automatic lines termination after bars closing
Whenever a group of bars (e.g. from nested loops, parallel execution, etc) finishes (which means all the bars are closed), cursor moves below the printed bars so that it's possible to print something after them.
Note: the behavior has also changed for the case when there are bars with positions specified, which run non-simultaneously. Every time a bar is created for a file where there are no bars running, the positions are numbered from the current cursor position. Could be a subject to fix later