feat: Raise error as soon as iter() is called if not wrapping any iterable
This is to replace the slightly more cryptic "None object is not iterable" you get if the tqdm is not wrapping any iterable (is in manual mode) with a more explicit error. Also, by calling this as soon as iter() is called and not when the iterator is first advanced, this kind of mistake is caught earlier.