Make `tqdm_asyncio.gather` interface consistent with the stdlib `asyncio.gather`
Intended as a drop-in replacement of the stdlib `asyncio.gather()`, it would be much reasonable if the function interface is consistent with that of `asyncio.gather()` as well.

It also ensures that our interface doesn't violate the least-surprising principle. It would have been surprising and frustrating for a user to wanting to use `tqdm_asyncio.gather()` as drop-in replacement to the stdlib `asyncio.gather()` but ending up being bitten by the trivial inconsistency.