Fix bug where the thread monitor checks miniter instead of mininterval
The comment above the condition of the monitoring thread that makes it
refresh the progress bar, states that the refreshing should happen only
if the mininterval of the instance is greater than 1 second, but instead
the miniter of the interval is checked.

Additionally, since we force set the miniter to 1 to bypass the user set
value, this condition could be true at most once, which caused the
monitoring thread to not refresh the progress bar according to
maxinterval.