docs: fix Build-Status badge
Both in the README (changed in .readme.rst) and in DEMO.ipynb

See the linked GitHub issue from the previous version of the badge:
https://github.com/badges/shields/issues/8671

In summary, these badges were changed in a breaking fashion such that
the badges were just linking to the github issue rather than showing the
data. Updating the url resolves this.

Also update the link of the badge to specify the workflow file rather
than the name, which matches the new behaviour of the badge.