tqdm.std: add update_to
Porting from dvc's Tqdm progress bar implementation.

https://github.com/iterative/dvc/blob/cf8195c829aa67425b87e4145bd33cae8584ee1e/dvc/progress.py#L114-L117

So, the implementation uses `self.update()` so that it refreshes based on `self.update()` logic.
Also, adds a way to update `total`.

This function is pretty convenient and useful even outside dvc. Take an example of fsspec's callback
that uses both `total` and `value` (i.e. current position) at once.

https://github.com/fsspec/filesystem_spec/blob/0ffe06cb767456b7c13904b57ec1c3ca60d53eae/fsspec/callbacks.py#L228

I see that there are mentions of `update_to` in docs and examples, which will likely have to change
as a result of this PR. So I'm happy to rename this to something else (eg: `set_position` etc.).

Related: #1264.