Improve robustness of IPykernel detection
This is not perfect but should help; in particular it is possible to
launch ipykernel without `'IPykernel'` being present in the
config, in the case of cloud based launcher (like
https://gateway-provisioners.readthedocs.io/en/latest/).

Note that with the ubiquitousness of tqdm, if this makes it easier for
you; I'm open to add a

```
try:
    import tqdm
    tqdm.in_ipykernel(...)
except ImportError:
    pass
```

Or anything similar if this helps detecting in which frontend you are
run from. I'm guessing this coudl help as sometime ipykernel is in a
notebook, but sometime in QT (like in spyder), and I guess it could help
decide what to do without user interactions.